+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 	Placeholder for the next version (add instead of version-number-headline below):
 	## __WORK IN PROGRESS__
 -->
+## 0.0.3 (2021-02-03)
+* Removed docs from NPM
+* Added loading-animation for model
+* Multiple Bugfixes
+
 ## 0.0.2 (2021-02-02)
 * Reordered some menues in VIS editor
 * Extended documentation